requirements.txt
redis==3.2.1
requests==2.27.1
python-dotenv>=0.18.0